Re-organize integration test
This simplifies the cmd folder and allows integration tests to be easily discovered
