+++ cmd/customer-svc/main.go
@@ -5,12 +5,12 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/albertowusuasare/customer-app/cmd/customer-svc/pkg"
 	"github.com/albertowusuasare/customer-app/internal/api"
+	"github.com/albertowusuasare/customer-app/internal/app"
 )
 
 func main() {
-	app := pkg.InmemApp()
+	app := app.Inmem()
 	port := ":5090"
 	log.Println(fmt.Sprintf("Starting server on port%s", port))
 	log.Fatal(http.ListenAndServe(port, api.Handler(app)))
